Cosmetic refactor
