+++ public/views/Register.js
@@ -1,4 +1,4 @@
-import DefaultView from "./view";
+import BaseView from "./BaseView";
 import Plate from "../components/plate/plate";
 import Text from "../components/text/text";
 import Form from "../components/form/form";
@@ -6,11 +6,7 @@ import FormField from "../components/form/__field/form__field";
 import FormError from "../components/form/__error/form__error";
 import BackendIntegrator from "../network";
 
-class Login extends DefaultView {
-    constructor() {
-        super();
-    }
-
+export default class Register extends BaseView {
     render() {
         const plate = new Plate();
         const text = new Text();
@@ -18,16 +14,9 @@ class Login extends DefaultView {
         const formField = new FormField();
         const formError = new FormError();
 
-        const dontHaveP = document.createElement("p");
-        const dontHaveA = document.createElement("a");
-        dontHaveP.innerText = "Don't have an account? ";
-        dontHaveA.href = "register";
-        dontHaveA.innerText = "Register";
-        dontHaveP.appendChild(dontHaveA);
-
-        const plateLogin = plate.render(
+        const plateRegister = plate.render(
             "plate__size-big",
-            text.render("text__align-center text__size-big", "h2", "Login"),
+            text.render("text__align-center text__size-big", "h2", "Register"),
             form.render(
                 "",
                 formField.render("login", "login", "form__field__size-big"),
@@ -39,17 +28,24 @@ class Login extends DefaultView {
                 ),
                 formError.render("Password incorrect"),
                 formField.render(
+                    "passwordRepeat",
+                    "repeat password",
+                    "form__field__size-big"
+                ),
+                formError.render("Passwords don't match"),
+                formField.render("email", "email", "form__field__size-big"),
+                formError.render("Email incorrect"),
+                formField.render(
                     "submit",
-                    "Login",
+                    "Register",
                     "button button__size-big button__color-lavender"
                 ),
                 formError.render("All is incorrect")
-            ),
-            dontHaveP
+            )
         );
 
         const platesWrapper = document.querySelector(".plates__wrapper");
-        platesWrapper.appendChild(plateLogin);
+        platesWrapper.appendChild(plateRegister);
         document.querySelector("form").addEventListener("submit", function(e) {
             e.preventDefault();
             console.log("submit");
@@ -57,13 +53,13 @@ class Login extends DefaultView {
                 .value;
             const passwordValue = document.querySelector("input[type=password]")
                 .value;
+            const emailValue = document.querySelector("input[type=email]")
+                .value;
             if (loginValue.match(/^\w{6,19}$/)) {
                 console.log("loginmatch");
                 const back = new BackendIntegrator();
-                back.login(loginValue, passwordValue);
+                back.register(loginValue, passwordValue, emailValue);
             }
         });
     }
 }
-
-export default Login;
