+++ public/www/index.html
@@ -4,8 +4,8 @@
         <meta charset="UTF-8" />
         <title>LeMMaS</title>
         <link rel="icon" type="image/png" href="../favicon.png" />
-        <script src="../bundle.js"></script>
     </head>
-
-    <body class="body"></body>
+    <body class="body">
+        <script src="../bundle.js"></script>
+    </body>
 </html>
